refactor(controller, service) : 수정, 삭제할 때 권한 체크

+++ src/main/java/com/woowacourse/zzinbros/user/service/UserService.java
@@ -2,8 +2,11 @@ package com.woowacourse.zzinbros.user.service;
 
 import com.woowacourse.zzinbros.user.domain.User;
 import com.woowacourse.zzinbros.user.domain.UserRepository;
+import com.woowacourse.zzinbros.user.domain.UserSession;
 import com.woowacourse.zzinbros.user.dto.UserRequestDto;
+import com.woowacourse.zzinbros.user.exception.NotValidUserException;
 import com.woowacourse.zzinbros.user.exception.UserDuplicatedException;
+import com.woowacourse.zzinbros.user.exception.UserLoginException;
 import com.woowacourse.zzinbros.user.exception.UserNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -24,14 +27,22 @@ public class UserService {
         throw new UserDuplicatedException();
     }
 
-    public User modify(Long id, UserRequestDto userRequestDto) {
+    public User modify(Long id, UserRequestDto userRequestDto, UserSession userSession) {
         User user = findUser(id);
-        user.update(userRequestDto.toEntity());
-        return user;
+        if (userSession.matchEmail(user)) {
+            user.update(userRequestDto.toEntity());
+            return user;
+        }
+        throw new NotValidUserException("수정할 수 없습니다");
     }
 
-    public void resign(long id) {
-        userRepository.deleteById(id);
+    public void resign(Long id, UserSession userSession) {
+        User user = findUser(id);
+        if (userSession.matchEmail(user)) {
+            userRepository.deleteById(id);
+            return;
+        }
+        throw new NotValidUserException("삭제할 수 없습니다");
     }
 
     public User findUserById(long id) {
@@ -40,6 +51,21 @@ public class UserService {
 
     private User findUser(long id) {
         return userRepository.findById(id)
+                .orElseThrow(UserNotFoundException::new);
+    }
+
+    public UserSession login(UserRequestDto userRequestDto) {
+        User user = findUserByEmail(userRequestDto.getEmail());
+        User requestUser = userRequestDto.toEntity();
+
+        if (user.isAuthor(requestUser)) {
+            return new UserSession(userRequestDto.getName(), userRequestDto.getEmail());
+        }
+        throw new UserLoginException();
+    }
+
+    private User findUserByEmail(String email) {
+        return userRepository.findByEmail(email)
                 .orElseThrow(UserNotFoundException::new);
     }
 }
